+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * An optional 'home button' has been added to the controls, allowing the user to zoom all of Switzerland into view.
 * `rokas-popup` and `rokas-teaser` components are made available to J-M-C integrators.
 * The option to show the map control buttons in a smaller size.
+* Add paginator functionality to popup and teaser.
 
 ### Changed
 * Breaking Change: The input field type ControlOptions has been split into UIOptions and InteractionOptions.